% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Efeitos.R
\name{Porcentagem}
\alias{Porcentagem}
\title{Cria grafico de porcentagem de efeito}
\usage{
Porcentagem(X, y, k, PC, t)
}
\arguments{
\item{X}{matrix}

\item{y}{vector}

\item{k}{numeric variable}

\item{PC}{numeric variable}

\item{t}{numeric variable}
}
\description{
Cria grafico de porcentagem de efeito
}
